Assert that shared/compose has a src to copy (WIP)
Because trying to copy from nil to an
InstrumentedUnpooledUnsafeNoCleanerDirectByteBuf led to a SIGSEGV, and
this is a lot faster/cleaner/easier than restarting my JVM every time I
add more logging to try to identify the problem.

Refactor-renamed some methods in client.hello to remove the redundant
'hello part of their names.