Start refactoring client-test setUp (WIP)
Don't want callers to need to copy/paste the server key initialization.

Then I realized that the test I thought I fixed in previous commit was a
false positive (as noted in the comments). All I managed last time was
to get it back to the failure it had been experiencing when I shifted
gears away from it a year ago.

Started switching the build-hello test (which is the next in that ns) to
use the modified setUp (raw-client) call. It also fails, in a way that
seems very suspicious (and probably unrelated to initial failure that's
still a false positive).