Get client/step-1 test past "server cookie" (WIP)
An array of all-zeros passes its litmus test and tries to "fork" a child
for messaging.

Which fails because the test is supplying nil for that spawn function.

This seems like it's probably a reasonable place to be, since I really
don't want the interaction to work that way.

I don't think. This seems like it needs to be its own fairly hefty
feature branch.