Start updating the way Client forks Message (WIP)
At least for the client's step-1 test.

There are a bunch of moving pieces involved here. Not the least of which
is the updated logging architecture (which needs to propagate further).

The client.state/fork call has to be failing immediately after it
triggers message/start! (assuming that ever returns). But I don't see
that failure anywhere.

Maybe because it happens on the agent thread.

Instead, I just see the messaging ioloop go into a fast timeout failure
loop.

Still, I'm pretty sure this qualifies as forward progress.