WIP: Fighting clojure.spec generation
Started refactoring the marshalling code into its own ns. Decided to try
to write a generative test to verify it. Not having any luck figuring
out how to get it to write the prefix byte array.

The truly obnoxious part about this is that the specs for those really
need to be magic constants based on what's in the official spec.

Decided to mark this as a milestone so I can get back to it in case
switching to that approach doesn't work any better (I *will* need to
generate specific-length byte arrays for most of the fields involved).

Bumped a couple of library dependency versions.

The circular nature of the dependencies between the root shared and
shared.constants finally tripped me up with I did the serialization
refactoring. So I started moving some of the pieces from
shared.constants into shared.specs, though the line between them is
terribly fuzzy.

Fixed a couple of test name typos under helpers-test that stemmed from
an incomplete renaming refactoring, but didn't show up until I restarted
my JVM.