Start digging into client/server interaction (WIP)
Started poking around in lots of nooks and crannies that have lots of
"FIXME: get back to this once messaging is done" comments/exceptions.

This part really revolves around deciding to set up a "real" interaction
test, like I have for messaging, and starting the refactoring to make it
possible.

Along with lots of comments about other pieces that need to get shifted
into place.