Switch server-test to use new factory ns (WIP)
And then started working toward getting the HELLO from the Client
created by that Factory over to the Server which is really the point to
this test.

Touched several files, but none of them matter much.

This version gets the HELLO to the Server. But then it fails because
it's trying to do something with a B] that only makes sense with a
ByteBuf. I blame recent decompose refactorings, but I don't think this
kind of test has ever been run before.