Get server cookie back to client, unit test (WIP)
Work through some problems left over in terms of decomposing the hello
packet on the server.

Then get it to build the cookie and forward that ByteBuf to the client.

My unit test has a glitch at this point because I need to translate that
into a byte array first.