Adjust message-tests for client ctor changes (WIP)
Actually, that part works fine.

But there are a lot of implications in the client pieces that fall out
from that refactoring. And that's broken a lot of code.

This isn't a terrible thing, since that code has never really worked
anyway. But this seemed like a good milestone point before I delve back
into that angle.