Refactored broadly, but shallowly (WIP)
Associate server name/ip and port with packets coming out of the
Client. This approach doesn't really cope with multiple potential
servers, but it shouldn't be awful to add them to the mix.

Refactor the ::server-name key to ::srvr-name for the sake of
consistency.

This version builds in my current REPL, but it might not survive a
refresh if I could run one.

Also made the executive decision that packets coming in from aleph will
have the :message value in a byte array rather than ByteBuf.