Fix problems w/ client's Hello (WIP)
Previous commit broke the packet's header by creating a string of the
byte array object rather than a byte array from the string version of
the prefix.

The byte array that got generated from that was 4 bytes too long. This
killed the "evenly divisible by 16" rule for the packet length, so the
server was just discarding it on arrival.

While I was debugging that, refactored away most of the duplication
among the various bit-twiddling pack! sizes. This approach still seems
dubious, but at least there's less copy/paste.