Step back. Examine why cookie->vouch breaks (WIP)
client is baked around a lot of C-level assumptions that I'm trying to
mimic using agents. Initiate is really the entry point to that: it's
where I'm trying to fork a child message ioloop in a way that doesn't
make sense for this to happen.

Really, I just rolled my "Don't do this" exception a couple of steps to
try to get a little closer to pinpointing how that interaction should be
handled.