Found/fixed bug that was blocking everything (WIP)
It turns out that s/keys has the unexpected (but documented)
functionality that all namespaced and spec'd keywords in the map being
checked will also be validated.

Which means that my loop-stopper function was getting called, whether or
not I included it in the map spec.

I also converted the server's active-client map from an Atom.

This gets me back to processing the incoming HELLO packet. Which fails,
because I've drastically rearranged the way current-client is handled
without updating any of the code that uses it.