Fix up spec checks for server's child spawner
There were pieces that were still checking for that in the spec, even
though it isn't kosher to do so.

Changed the key name to indicate that it's something that triggers
side-effects and needs to be handled with kid gloves.

The same thing happens on the client side, but that's an issue for a
different day.

The server responds to the client's HELLO again. Then the client goes
into a fast-spin failure without an obvious reason that I've spotted
yet.

That seems like a good stopping point for the night.