Find latest failure: client/server handshake test
It boils down to the crypto namespace failing to create a new lock file
for loading the nonce.

Actually finding this involved quite a few changes.

Fixed a bug in one of my exception logging arities that was dropping the
actual exception.

Fixed a problem in the client Initiate code that was dropping exception
logs completely.

Updated the alt message ioloop scheduling code. What it was doing at the
bottom of the cond-> didn't really make any sense. It was sending back a
nil next-time under some circumstances (it may still be, but at least
now it should be easier to figure out why).

Added some extra error handling/logging around the actual problem to try
to figure out why it's doing what it is. That managed to break things in
a different way: the new exception is just getting swallowed.

Refactored my "other" non-STDOUT loggers to use the same method to
format log messages instead of rolling their own incorrectly.

Converted the client/server "instances" in that test to use separate
file loggers so I can look at what they're doing and try to make sense
of it separately.

Cleaned up some comment cruft about the previous roadblock.