Call test-factory client/start! in future (WIP)
Now it blocks while waiting for a response from the server. Which will
never happen if the thread itself is blocked waiting for the start! call
to return so I can pull the packet it's sending.

Got annoyed by the fact that my host/port are still getting set
incorrectly when the client sends them. Worked a little on sorting that
out and getting a real IP address into the client.

Broke my CIDER connection by running a refresh that failed. When I got
everything reconnected, trying to construct a Server failed because the
spec still requires a child instance.

Seemed like a reasonable stopping point.