Fix the NPE from this in servers-polled (WIP)
The problem stemmed from a keyword with the wrong namespace when I set
up the deferred chain. It looked like the plumbing was correct because
of the schizophrenic logging situation.

This version gets down to a newly-added NotImplemented exception: I
really need to consolidate the agent with the state value that I just
returned.

Or just ditch the Agent completely.