Start debugging Initiate creation (WIP)
Really just digging into the code that pulls the initial Message from
the forked [Client] child to combine that with the Vouch built from the
server's Cookie.

I managed to trim back one piece that was relying on an Agent, but the
rest of the code involved here still needs a lot of work.