Poke around a bit more among Client (WIP)
Spotted a couple of minor problems, then ran across another big one: as
things stand, we're getting something back from a server that looks like
a Cookie. We're checking the basic shape, accepting it (and thus the
server from which it came), and then later trying to open its associated
crypto box.

This may turn out to be minor, but I wanted to get a milestone commit
into place.