Debug previous commit (WIP)
Fixed a couple of problems, like trying to fork the child with a
startable ioloop description rather than the actual io-handle that
provides access to a running message loop.

But the actual Initiate packet still isn't showing up.

And there's an issue with one of the main message ioloops stopping.