Fixed problem with nil log-state in safe-nonce WIP
The real issue was that nothing in crypto had started using the new
logging system, but I was acting as though this had.

That freed things up to fix several fairly minor problems.

Like function rename refactorings that didn't propagate everywhere.

Or missing parameters in function calls.

Or checking for an agent error before trying to get it to take the next
steps.

This gets the Initiate packet out of the client. Now the tests time out
waiting for a Message response from the server.