Gank the client agent from a few more places (WIP)
- Query for the current timeout (which should probably just go away
  completely)
- Forking the child
- Building/sending the Initiate packet
- state/final-wait

I had to replace that agent with extra deferreds, and move back the
places where I send it messages to update its state. This part is
finicky enough that I've managed to break upstream pieces: I'm back to
not even being able to pull the hello packet from the client.