Decided to convert client start to deferred (WIP)
As opposed to using an agent as the foundation the way I started.

There's a little more debug logging in here, and I've spotted/fixed at
least one silly arity exception as I started converting the error
handling in the handshake unit test.

There's also a test for the way one of the constant namespaces used to
pull specific message lengths. The function it's based around went away
a long time ago. Decided to just punt on this failure for now, since I
have bigger fish to fry.