Remove dependency to io-handle in child->
It was really just there because I needed to keep logs synchronized via
an atom.

Since the logger can handle that (refactored out a function to make it
quick/easy to call), that was able to go away.

It was a good simplification, and cleared the next log jam.

Now we get back to pulling the server's message. I'm pretty sure that's
failing because I haven't written it yet.