Track down garbage Initiate packet souce (WIP)
This is really just minor little logging tweaks, with some added
comments and TODO items.

But I've managed to track down the source of the garbage message that
the server receives. And thrown a lot of exceptions around missing
pieces in client.initiate that I never got around to translating.