Get the client to build real Initiate packet (WIP)
It's going through the motions, at least. Mostly.

There are at least 3 fairly nasty broken pieces in here:

1. The server handshake-test just throws a StackOverflow error that
looks like it's caused by nrepl trying to print a recursive data
structure.
2. The Initiate packet it sends to the server has the packet envelope as
its :message member.
3. The :message member inside the nested envelope is a log/state instead
of a byte-array
4. The Initiate Packet builder code tries to get the server name from a
key that doesn't match where the value is found in that map. This isn't
as straightforward as it should be, because the key that's being used
doesn't make sense.

Still, this is forward progress. There are enough changes in here that I
wanted to get it logged in as a milestone.