Build crypto-box return in place (WIP)
There's a lot going on here, even though the changes really don't seem
to amount to much.

Actually, build-crypto-box was already building the [B that it
returns. But I was passing in a dst [B parameter and pretending that
that's where the data was getting shoved.

The original approach might not have been a premature optimization, but
it was fraught with peril, and I didn't translate it correctly anyway.

That baby-step is really just half the picture.

Building that crypto box fails because it's missing the inner-in-nonce
field. This raises an Assertion error.

The client still sends a broken packet (with the previously described
nested map) to the server.

And then the test turns into a false positive.

The mysterious StackOverflow error that's been plaguing me has
disappeared.