Fix most Client Initiate keyword mismatches (WIP)
Get it right on the verge of sending its broken packet to the Server.

This is really about knocking down my problems from the bottom up. I've
left one in place as an obvious marker.

The difficulty in the next step is that it just flips back over to a
StackOverflow exception that I still haven't had any luck deciphering.

There's also still the question of "How is this sending garbage to the
Server?"

I don't really want to clean up the last bits and pieces on this end
until I have a firm grasp on what's going on with those problems.