Fixed 1 double-envelope, found another
client.state was taking the byte-array(?) returned by
build-initiate-packet, wrapping it in a network packet envelope, and
then forwarding it to do-network-send as the message parameter which
would get wrapped into the network packet envelope.

That was one problem, which should be fixed now.

The second problem is broader in scope. At least I think I've identified
it now.

The basic idea there is that I'm taking the network envelope that
contains the Cookie Packet from the Server and trying to treat it as the
initial outgoing Message block being sent by the Client's Child.

I still need to sort through all the balls that got dropped here.

I also narrowed the scope of the client.initiate/build-initiate-interior
parameters. Instead of passing in a full ::state/state, I'm specifically
just selecting the keys from the map that it actually needs to do its
job.

The jury's still out on that one.