I was playing both ends against the middle (WIP)
I was forking the client's child, and then immediately trying to
build/send the Initiate packet.

And I was trying to use the just-arrived cookie message envelope as the
byte-array initial Message block from the Child.

I think I sorted out the basic problem by allowing the child-> callback
to do its job of
a) calling the packet-builder function and
b) calling do-send-vouch

I'm not sure about this. My filter function refuses the bytes that just
arrived from the child. And then I wind up with that ridiculous
StackOverflow with no meaningful context that I can find.

And then I'm not successfully stopping the Message ioloop, so we keep
failing to build that packet.

Still, this seems like really nice forward progress.