Stop client child, w/out logging (WIP)
This change looks a lot bigger than it really is.

Trying to figure out where client/stop! is going wrong.

This is almost all about logging and exception handling.

I did spot/fix a failure in the way I copy/pasted the logic flow. If you
tried to stop a Client that had no log-state, it wasn't calling the
function to actually stop the Child ioloop.

I had high hopes that fixing this would solve the issue where that
ioloop isn't getting stopped.

That was before I realized that the StackOverflow is killing my test
just before I call this function, so this change is largely meaningless
for the task/problem at hand.

Still, squashing bugs is always a positive thing.