Fix client's incoming Vouch filter
When the client received bytes from its child, I was validating them as
an Initiate packet.

Validating them as the message contentportion of that packet wouldn't be
a terrible idea, but this was comparing the length against something
completely unrelated. Maybe the length of the Initiate header, before
the actual Message cryptobox?

Whichever. For now, that check is gone.