Find/fix cause of stack overflow w/out context WIP
I was doing a nested send and returning the internal send's return
value. This left the agent with itself as its state.

Now I seem to be successfully stopping the client's child.

The test still times out while I'm waiting for the Client's Initiate
Packet, but this is a problem that's been haunting me for days.