Hedge optimism re: start from last commit (WIP)
It turns out that just getting to the point of telling the Client Child
to stop! doesn't mean that it was actually doing so.

Convert the message/action-trigger to a lexical closure around a
promise, with an interruption function, to give myself the opportunity
to just stop things manually at that level so I don't have to restart
the JVM every time I get tired of all the log spam from orphaned message
loops.

And also work on debug logging to try to sort through why/where it's
going wrong.

At a different layer, I can see when/where my handshake test is timing
out. The Client fails to build the Initiate packet because it's missing
the inner-i-nonce.