Eliminate half of last references to client agent
Well, maybe 3/8 instead of 4/9. I've counted it both ways.

What's left is pretty thoroughly isolated in a function that's tangled
up enough that I'm not tackling it late at night.

Moved the hello/do-build-hello down to the public section and renamed it
to do-build-packet in the process.

Fixed the spec for state/->message-exchange-mode. That thought it was
another reference to the client wrapper, but I'd already refactored the
function parameters.

Added a spec for its return type.