Track down missing Initiate pieces (WIP)
Found the Client's missing inner-i-nonce: I needed to build the inner
vouch before calling fork! so child-> would be able to access it.

Building the cryptobox that contains that (plus the initial Child
Message) failed because it was missing the shared secrets. Mostly the
short/short keypair which actually encrypts that box.

This version still fails because of a spare nil during Composition.

But it's a big/messy enough change that gave me enough trouble that I'm
committing it  so I can get back here if need be.