Fix Client Vouch filtering (WIP)
What was being filtered was a bit tangled up. I was trying to filter the
full Initiate packet length based on the legal length of the Message
payload it can possibly contain.

And then I was trying to check that the length is evenly divisible by 16
by calling quot rather than rem.

And now I'm trying to convert what I'm about to send into a string, but
I don't have a good ByteStreams conversion for that.

It isn't a lot of forward progress, but it's heading in the proper
direction.