Give up on byte-streams ByteBuf converter (WIP)
For now, at least, just do the conversion manually and call print-bytes
on the byte-array I read the old-fashioned way.

This version gets the Initiate packet to the Server. It fails there
because the Server's been relying on receiving byte-arrays, and this
payload is arriving as a ByteBuf.

I don't want to mess with that for now, since I think that's what I
expect to receive, ultimately.