Fix several minor nuisance issues
Got rid of a duplicate child-spawner! spec in client.state. Added a note
to the server equivalent that it should also switch to using the shared
version.

Rearrange the root message interrupt! pieces for the action-trigger so
they're all clumped together rather than scattered at top and bottom.

Cleaned up some of the noise in the :require section of the ns
declaration.

Fixed the serialization-test ns to match its folder location, so it
won't keep me from running (refresh)