Debugging step-1 test failures (WIP)
The Server sends a bogus Cookie that I can't decrypt. I'm sending back
an empty Initiate Packet in response. I should just be silently
discarding it.

This was all about tracing through the code to try to figure out where
that's going wrong.

Finally refactored set-up-server-polling! from client to hello. This is
completely and totally untested. It's just something that's been bugging
me for a really long time, and this seemed like a good chance to sort it
out.

Even though, now that I've done it, all the callbacks I had to supply
seem pretty awkward.

Added a README to try to help me keep track of the overall flow
chart. It's painfully twisty.