Refactor cookie/received-response! to callback WIP
Instead of calling success! on a deferred.

Note that I'm still passing in the same deferred as a parameter. I just
happened to get lucky: Deferred implements IFn. So calling it works the
same as calling success! on it.

This basic change was simple/easy enough that I moved on to the next
step, which is to convert that deferred to a real callback that I
control.

This turned out to be a can of worms.

I don't think that any of that work made its way into this commit. But I
wouldn't be surprised if it did.