Refactor hello polling loop to recurse and chain
Main point was switching it to a deferred/chain. That took away my
ability to recur.

This allowed me to refactor the logic for moving on to the next server
after a timeout into a top-level stand-alone function, eliminating some
duplicate code.

I also toyed with the actual timeout numbers. Switch to generating them
the same way as the reference implementation.