Remove completion from hello polling loop (WIP)
It's a deferred that I've been using to set up the call to start
building the Initiate packet.

Eliminating it gets rid of a lot of spooky action at a distance that was
making the overall control-flow very difficult for me to follow.

At the same time, now my client's Initiate packets aren't getting
built.