Refactor polling-loop's cookie-sent to top (WIP)
i.e. Move it from an inline functional clojure to a top-level function.

This is a first step in an attempt to make the pieces small enough to
help figure out where the Initiate response is disappearing (yet again).

Fixed a bug in an exception handler where I was trying to call
flush-logs! without a logger.

Fixed another in a different one where I was just building the new log
message and discarding it.