Found disappearing server-cookie
Converted the spec for the black box to something a bit more restrictive
than any?

cookie/servers-polled no longer returns a tuple of (::log/state and the
deferrable which would eventually contain the ::state/state). Now it
just returns the modified state. Remove the intermediate step in the
deferred chain in client/start! that was juggling back to this.

In server-test/handshake-test, the Initiate packet heading to the server
has sometimes been a byte array. Others, it's a ByteBuf. Just always do
the conversion to be sure.