Revisit errors in cookie/received-response! (WIP)
When I threw an exception in the middle of this, it was basically
stopping the client, but then it still forked and sent out an empty
Initiate Packet.

Fixed some problems that swirled around "What happens next?" that mostly
amounted to Arity exceptions that got introduced when I rearranged
most of the Hello/Cookie polling loop to reduce callbacks.

(There were also problems with the arguments and implementation of
pick-next-timeout).

Refactored the hello spec for the cookie-sent-callback to the top
level. Although, really, it should just use the same spec as the actual
function in the cookie ns.

Brokenness:

The recursion to poll the next server is doing something wrong enough
that it seems to unblock the deferred chain.

Left it there.