Fix conflicting server-ips for Hello polling
Most things were using a lazy seq that I was passing around as an extra
parameter. But possibly-recurse was using the original value that I set
up with the original list instead, so it was failing after the first
try (since my handshake test only has a single server in the list).

Flush logs when exception reaches end of deferred chain in client/start!
I was just returning them, but that return value gets ignored.

Added an alias for byte arrays to make it more convenient to convert
other types to them.

Fixed some arity exceptions in my server handshake test where it was
calling ex-info with nothing but the message argument.

This version looks like the client is sending Hello twice and getting
back two Cookies.