Get Initiate Packet back to server
This really deserved more than a single commit.

Verified that, if an incoming Cookie doesn't work, the Client will move
on to send a Hello to the next Server in its list.

Cleaned up some functions in client.initiate that have been deprecated
for quite a while now.

Track down a log message in client.state/child-> that was swallowing the
Initiate Packet before I could actually send it. Switched that to just
convert the ByteBuf to a byte array.

Now that Initiate Packet gets back around to the Server.

And then the Server discards it immediately. It looks like it does so
because the string header it expects is actually a byte array.

That's a problem for a different branch.