Switch to a shared spec for time descriptions
I'm tempted to narrow it down based on time scale, and I keep wanting to
force it to be integral, but I think this is a pretty reasonable
compromise.

Add a note that the client README is badly out of date.

And a comment about the way client.hello/do-build-packet works: it
really should be handled differently, but that's a task for a different
feature branch.