Kick the tires on removing ::working-nonce
Well, pulling the concept out of the main top-level state, anyway.

Starting with the server cookie.

This looks like a bigger change than it is, due to some whitespace
changes.

The only thing in here that's even vaguely substantial is adding a
wrapper function around crypto/do-safe-nonce so callers can just call
get. That's deceptive since it *is* stateful, but it's also safer
because it doesn't involve modifying buffers in place.

Everything in life's a trade-off.