Refactor away server.cookie working-nonce (WIP)
Focused on prepare-packet!

The core of what's going on here swaps out a bunch of manual byte
manipulation in favor of crypto/build-crypto-box.

This has been a FIXME task for quite some time now.

This version is broken: when the Client tries to decrypt, it triggers a
NPE. Shortly after that (quite probably related), there's also an
AssertionError about a missing nonce. I haven't really looked at that
side of things yet.

This turned up a few other pieces of low-hanging fruit.

Switched bit-twiddling/copy-bytes! to call (bytes) rather than
type-hinting (as recommended in the clojure docs about primitive
JVM interop).

Fixed a log call in the crypto ns that didn't include the log-state.

Updated logging/exception to cope with cases where it doesn't get an
exception in the correct spot (this was caused by that missing log-state
parameter).