Make client cookie handling more robust
When a Cookie arrives, verify that we really did decrypt it before
trying to deserialize it.

When it fails (which it does now, due to ways I broke the server in
previous commits), have possibly-recurse call do-polling-loop with the
proper number of arguments.

Also, just recurse. I've given up on using trampoline for this, for
now. So returning a partial function wasn't working at all.