Verify that refactored inner-cookie is wrong (WIP)
Building it doesn't match the original. I'm doing something wrong with
the test, though.

New version produces 96 bytes, which is expected. Old version produces
112, which seems wrong.

I think there's something wonky with the overlap here, though. Like
maybe the first 16 bytes of the original are garbage that filled in the
0 padding.

But I'm out of time for working on this tonight.