Get new approach to inner cookie working (WIP)
Encrypting the crypto box wrapper around it still fails. I'm pretty sure
this is due to half-complete changes to the way I'm handling the
nonce. The bulk of the changes here stem from trying to find and verify
that.

Eliminated at least one of the really sticky places where I'm still
using the side-effects logger.