Fix broken cookie construction
It all boiled down to replacing the working nonce. I was still using its
bytes to populate the plain-text nonce in the packet wrapper, even
though I was using the safe nonce to actually do the encryption.

Fixed at least 1 other bug along the way. I was setting up a retry to
time out in nanoseconds, when it uses milliseconds.

This version is back to failing in the "expected" way: the server fails
to respond to the client's initial message.