Debug why server drops Initiate packets (WIP)
There are at least 3 different threads of changes worming through here.

Start making the conversion to functional logging. Had to do this
because we call a function in crypto that has also made that switch.

Unfortunately, it makes debugging the other problems more challenging.

Started refactoring a do-fork-child function out of the middle of
do-handle. Which used to be named handle!

And then I'm trying to debug why do-handle fails silently.