Refactor low-hanging server fruit (WIP)
These pieces all turned up while I continued debugging the disappearing
Initiate packet.

Moved the root server's handle-hello! function into the hello ns and
renamed it to do-handle. This has been on my TODO list for ages.

Renamed its handle-message! and handle-incoming! to do- variants, to
reflect that their return values matter (in addition to their
side-effects. I'm sorry about that).

Pulled some now-pointless log-cleaning code from server.helpers.

Added a lying spec to server.state to indicate that a bunch of the
return values from these changed functions are really meant to be deltas
that get merged into the existing state.

Add a couple of specs and comments to shared.serialization.

This probably looks like a fairly hefty change in terms of lines of code
altered, but it really doesn't amount to all that much.