Start refactoring server.initiate spaghetti (WIP)
Extracted a couple of fairly obvious functions from do-handle.

Looks like a bigger change than it really is, because I also got rid of
a pointless outer or block that returned an empty map for nil changes,
so it updated all the indentation below that.

Probably should have just stuck that into its own commit to make it
glaringly obvious.

Added a couple of really minor changes to server.state. Have it throw an
exception rather than just log an error if someone tries to set up a
client that's missing either half of the shared key. Added a spec for
the return value of find-client.

Haven't had time to run tests yet.