Refactor server.initiate decryption (WIP)
This is another one that looks a lot bigger than it really is.

This is mostly reworking decrypt-initiate-vouch and decrypt-inner-vouch.

They're more succinct, and neither modifies a dst parameter in-place any
more.

Clarify the problems I have with verify-client-pk-in-vouch. I've
verified that the keys it compares are obviously different because one
is still encrypted. And that I need to verify why bytes= doesn't
recognize that they're obviously different.

Functional logging is starting to trickle up from the bottom.

Converted crypto/secret-box and secret-unbox to build their own dst
return values rather than modifying a parameter in-place.

dissoc the ::index used during reduce from
shared.serialization/decompose-array.