Find the real culprit behind last night's mystery
The crazy type error I was seeing was actually the assert failure
showing up right after a nil.

The root cause of this is that I never seem to have gotten around to
actually setting up/configuring a new client connection definition. So
that's nil when I tried to interact with it.