Sort through options for server clients (WIP)
Adjust the function for coping with follow-up Initiate packets. The
basic idea still throws a NotImplementedException, but I think the docs
about what could/should happen make sense.

If there's a reason to forward along a message to an existing child, the
relevant code should be the same, whether that message is going to a new
one or is coming from a new Initiate packet.

Refactored the code for setting up a new client into its own function.

It still doesn't work, but it least I've established some isolation
around it.