Clean up based on visual scan of branch (WIP)
A lot of this is looking for TODO items and either converting them to
RuntimeExceptions for things that will take longer to implement or just
implementing them if the change is straight-forward and obvious.

Completely and totally untested. I can guarantee it is less correct than
previous commit, if only because of those TODO exceptions that I've injected.