Dig deeper into server's pk-triad failures (WIP)
Got a bit derailed by logging issues.

An exception is getting thrown somewhere that discards logs.

The top-level logging atom is getting set to nil sometime before that.

I think the two are unrelated, but they combine to make life difficult.

This is another of those "out of time for the evening" commits.