Fix "decryption" init server failure (WIP)
The actual problem was a misplaced paren in the if check for
success. The then clause was just getting discarded, and the else clause
was always getting executed. So it was literally a false negative.

Add a log message to unit test from previous commit so we can see the
the shared secrets on failure to verify that the comparison check isn't
a false negative (because I don't trust my bytes= translation).

Added a "Box Opened" log message to open-after to verify the
success. Log the return code on failure, though I don't think it has a
meaning documented anywhere.

It still fails to build a new client, but at least this is no longer the
reason why.