Rewrite basic child fork from server (WIP)
The basic idea is in place. And looks like it will be a huge improvement
all around.

It still needs quite a bit of tracing the tendrils in both directions to

a) pick up the changed incoming fork! parameters
b) cope with the changes to what happens to the outgoing ioloop