Fix the main problem identified in previous commit
Figured out where the message-loop-name was disappearing (I didn't
have it nested deeply enough in the tree).

Server really needs to set up the child's message-loop-name when it
spawns the child. Add a base-name member it can use to build that name
when it's time.

This version gets the server's child fork down to the point where it
should send a message back. And the related NotImplemented exception
that I added as a reminder that I really need to consolidate this
piece also.