Start converting server.initiate to new child (WIP)
i.e. switch the way it interacts with its child to the way I think I
have intended as the new shared approach.

It's been long enough since I wrote this code that I've gotten fuzzy
about the details of actually using it, and it looks like I've punted
on the client implementation, so I don't have an example to follow.

So I'm starting to break new ground, which is fun.

This approach most definitely does not work: it throws a NPE when it
tries to decrypt the Message portion of the Initiate Packet sent from
the Client.