Debug broken idea behind previous commit (WIP)
Do not need to decrypt the Message inside a Client's Initiate Packet
because it's already been decrypted.

This circles back (again) to the basic fact that Message Packets are
very distinct from the stream of Messages that they (along with Client
Initiate Packets) happen to contain.

Just need to find the Message and forward it along to the child.

This is a very broken attempt to do just that.

The problem is a little more clear than it was in the previous commit,
but it's probably further from the solution.